media_store/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "media-store"
version = "0.1.0"
description = "FastAPI-based microservice for managing media entities with metadata extraction, versioning, and duplicate detection"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "sqlalchemy-continuum>=1.4.0",
    "clmediakit @ git+ssh://git@github.com/asarangaram/clmediakit.git",
    "pytest>=7.4.0",
    "httpx>=0.24.0",
    "python-jose[cryptography]>=3.3.0",
]

[tool.setuptools.packages.find]
include = ["entity*"]
exclude = ["tests*", "alembic*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["entity"]
omit = ["tests/*", "*/conftest.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

